# architecture/openapi.yaml
openapi: 3.0.3
info:
  title: AI CREAT - API Specification
  description: API for the AI CREAT application, enabling users to upload, process, and generate creative assets, and allowing admins to manage the platform's rules and templates.
  version: 1.1.0
servers:
  - url: /api/v1
    description: Main API endpoint
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string, example: "naomi" }
                password: { type: string, format: password, example: "password123" }
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: User Logout
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Successfully logged out.
...
  /generate:
    post:
      tags: [Generation]
      summary: Start a New Generation Job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerationRequest'
      responses:
        '202':
          description: Generation job accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId: { type: string, format: uuid }
  /generate/{jobId}/status:
    get:
      tags: [Generation]
      summary: Get Generation Job Status (for polling)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Current status of the job.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [pending, processing, completed, failed] }
                  progress: { type: integer, example: 75 }
  /generate/{jobId}/results:
    get:
      tags: [Generation]
      summary: Get Generation Job Results
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: An object containing generated assets grouped by platform.
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/GeneratedAsset'
                example:
                  Instagram: [{...}, {...}]
                  Facebook: [{...}]
  /generated-assets/{assetId}:
    get:
      tags: [Generation]
      summary: Get a single Generated Asset
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          description: The requested generated asset.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedAsset'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags: [Generation]
      summary: Apply Manual Edits to a Generated Asset
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssetId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                edits:
                  type: object
                  properties:
                    crop: { type: object, properties: { x: {type: number}, y: {type: number}, width: {type: number}, height: {type: number} } }
                    saturation: { type: number, minimum: 0, maximum: 2 }
                    textOverlays:
                      type: array
                      items:
                        $ref: '#/components/schemas/TextOverlay'
                    logoOverlay:
                      $ref: '#/components/schemas/LogoOverlay'
      responses:
        '200':
          description: Asset updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedAsset'
  /download:
    post:
      tags: [Generation]
      summary: Get Download URL for Assets
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [assetIds, format, quality, grouping]
              properties:
                assetIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                format: { type: string, enum: [jpeg, png] }
                quality: { type: string, enum: [high, medium, low] }
                grouping: { type: string, enum: [individual, batch, category] }
      responses:
        '200':
          description: A pre-signed URL for downloading the assets.
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloadUrl: { type: string, format: uri }

  # -------------------------------------------
  # Admin - Platforms & Formats
  # -------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    JobId:
      name: jobId
      in: path
      required: true
      schema: { type: string, format: uuid }
    AssetId:
      name: assetId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
...
    GeneratedAsset:
      type: object
      properties:
        id: { type: string, format: uuid }
        originalAssetId: { type: string, format: uuid }
        filename: { type: string }
        assetUrl: { type: string, format: uri }
        platformName: { type: string, nullable: true }
        formatName: { type: string }
        dimensions: { type: object, properties: { width: {type: integer}, height: {type: integer} } }
        isNsfw: { type: boolean }
    GenerationRequest:
      type: object
      properties:
        projectId: { type: string, format: uuid }
        formatIds:
          type: array
          items: { type: string, format: uuid }
        customResizes:
          type: array
          items:
            type: object
            properties:
              width: { type: integer }
              height: { type: integer }
...
    TextOverlay:
      type: object
      properties:
        content: { type: string }
        textStyleSetId: { type: string, format: uuid, description: "ID of a predefined admin text style set" }
        styleType: { type: string, enum: [title, subtitle, content] }
        position: { type: object, properties: { x: {type: number}, y: {type: number} } }
    LogoOverlay:
      type: object
      properties:
        logoUrl: { type: string, format: uri, description: "URL to an uploaded logo asset" }
        position: { type: object, properties: { x: {type: number}, y: {type: number} } }
        size: { type: number, description: "Size as a percentage of the canvas width" }
...